check owner_id before return

+++ app/routers/post.py
@@ -6,7 +6,7 @@ from sqlalchemy.orm import Session
 from .. import models
 from ..oauth2 import get_current_user
 from ..database import get_db
-from ..schemas import PostBase,PostCreate,PostUpdate,PostResponse,UserOut
+from ..schemas import PostBase,PostCreate,PostUpdate,PostResponse,UserOut,PostCreated
 
 router = APIRouter(prefix = '/post',tags = ['posts']) # create router object
 
@@ -14,14 +14,17 @@ router = APIRouter(prefix = '/post',tags = ['posts']) # create router object
 def get_posts(db:Session= Depends(get_db),token: str = Header('Authentication')): # Pydantic format
 	#'clear', 'copy', 'fromkeys', 'get', 'items', 'keys', 'move_to_end', 'pop', 'popitem', 'setdefault', 'update', 'values'
 	current_user = get_current_user(token)
-	print(current_user)
+	# print(current_user)
 	posts = db.query(models.Post)
 	return posts.all() #{"data":posts.all()}
 
 @router.post("/",status_code = status.HTTP_201_CREATED,response_model = PostResponse)
 def create_post(post:PostCreate,db:Session = Depends(get_db),token: str = Header('Authentication')): #,user_id: int = Depends(get_current_user)):
 	current_user = get_current_user(token) # verify user login by token
-	new_post = models.Post(**post.dict()) # unpackage form 
+	# print(current_user)
+	# post = post.dict() # convert to dict
+	# post['owner_id'] = current_user['user_id']
+	new_post = models.Post(owner_id = current_user['user_id'], **post.dict()) # unpackage form 
 	db.add(new_post) # add new row
 	db.commit() # commit to save it to database
 	db.refresh(new_post)
@@ -35,6 +38,11 @@ def get_post(id: int,db: Session = Depends(get_db),token: str = Header('Authenti
 	if not post:
 		raise HTTPException(status_code = status.HTTP_404_NOT_FOUND,
 							detail = f"post with id {id} not found")
+
+	if post.owner_id != current_user['user_id']:
+			raise HTTPException(status_code = status.HTTP_401_UNAUTHORIZED,
+							detail = f"this post is not belong to this user")
+
 	return post #{"post_detail": post}
 
 @router.delete("/{id}",status_code = status.HTTP_204_NO_CONTENT)
@@ -48,31 +56,37 @@ def delete_post(id:int,db:Session = Depends(get_db),token: str = Header('Authent
 	else:
 		# db.delete(post)
 		# db.commit()
-		post_query.delete(synchronize_session =False)
-		db.commit()
+		if post.owner_id != current_user['user_id']:
+			raise HTTPException(status_code = status.HTTP_401_UNAUTHORIZED,
+							detail = f"this post is not belong to this user")
+		else:
+			post_query.delete(synchronize_session =False)
+			db.commit()
 
 	return Response(status_code = status.HTTP_204_NO_CONTENT)
+
+@router.get("/mine/",response_model = List[PostCreated]) # '/mine/' NOT '/mine'
+def my_posts(db:Session= Depends(get_db),token: str = Header('Authentication')):
+	current_user = get_current_user(token)
+	my_posts = db.query(models.Post).filter(models.Post.owner_id == current_user['user_id']).all()
+	return my_posts
 
 @router.put("/{id}",response_model = PostResponse)
 def update_post(id:int,update_post:PostBase,db:Session = Depends(get_db),token: str = Header('Authentication')): #,data:dict=Body(...) must be in the last of declaration
 	current_user = get_current_user(token) # verify user login by token
 	post_query = db.query(models.Post).filter(models.Post.id == id)
 	post = post_query.first()
-	print(post.title)
-
 	if not post:
 		raise HTTPException(status_code = status.HTTP_404_NOT_FOUND,
 							detail = f"post with id {id} not found")
-	
-	# post.update({'title':update_post.dict()['title'],
-	# 			'content':update_post.dict()['content'],
-	# 			'published':update_post.dict()['published']
-	# 			},synchronize_session =False)
-	
+
+	if post.owner_id != current_user['user_id']:
+			raise HTTPException(status_code = status.HTTP_401_UNAUTHORIZED,
+							detail = f"this post is not belong to this user")
+
 	post_query.update(update_post.dict(),synchronize_session =False)
 	db.commit()
-
-	return post_query.first() #{"updated_post": post_query.first(),"message":"post updated!"}
+	return post_query.first()
 
 @router.patch("/{id}",response_model = PostResponse)
 def update_field(id:int,response:Response,data:dict=Body(...),db: Session= Depends(get_db),token: str = Header('Authentication')): #,data:dict=Body(...) must be in the last of declaration
@@ -82,6 +96,11 @@ def update_field(id:int,response:Response,data:dict=Body(...),db: Session= Depen
 	if not post:
 		raise HTTPException(status_code = status.HTTP_404_NOT_FOUND,
 							detail = f"post with id {id} not found")
+
+	if post.owner_id != current_user['user_id']:
+			raise HTTPException(status_code = status.HTTP_401_UNAUTHORIZED,
+							detail = f"this post is not belong to this user")
+
 	# update
 	if 'title' in data.keys():
 		post.title = data['title']
